# control/api/internal/handlers/placements.go
package handlers

import (
	"net/http"
	"strconv"

	"github.com/gin-gonic/gin"
	"github.com/inscenium/inscenium/control/api/internal/db"
	"github.com/sirupsen/logrus"
)

// PlacementHandler handles placement-related requests
type PlacementHandler struct {
	db *db.DB
}

// NewPlacementHandler creates a new placement handler
func NewPlacementHandler(database *db.DB) *PlacementHandler {
	return &PlacementHandler{db: database}
}

// PlacementOpportunity represents a placement opportunity (simplified)
type PlacementOpportunity struct {
	ID          string  `json:"id"`
	TitleID     string  `json:"title_id"`
	ShotID      string  `json:"shot_id"`
	StartTime   float64 `json:"start_time"`
	EndTime     float64 `json:"end_time"`
	PRSScore    float64 `json:"prs_score"`
	SurfaceType string  `json:"surface_type"`
	CreatedAt   string  `json:"created_at"`
}

// ListOpportunities handles GET /opportunities
func (h *PlacementHandler) ListOpportunities(c *gin.Context) {
	titleID := c.Query("title_id")
	minPRSStr := c.DefaultQuery("min_prs", "0")
	
	minPRS, err := strconv.ParseFloat(minPRSStr, 64)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid min_prs parameter"})
		return
	}

	logrus.WithFields(logrus.Fields{
		"title_id": titleID,
		"min_prs":  minPRS,
	}).Info("Listing placement opportunities")

	// TODO: Implement actual database query
	// For now, return mock data
	opportunities := []PlacementOpportunity{
		{
			ID:          "surface_001",
			TitleID:     titleID,
			ShotID:      "shot_001", 
			StartTime:   5.2,
			EndTime:     12.8,
			PRSScore:    87.5,
			SurfaceType: "wall",
			CreatedAt:   "2024-01-15T10:30:00Z",
		},
		{
			ID:          "surface_002",
			TitleID:     titleID,
			ShotID:      "shot_002",
			StartTime:   15.1,
			EndTime:     23.4,
			PRSScore:    92.1,
			SurfaceType: "table",
			CreatedAt:   "2024-01-15T10:30:00Z",
		},
	}

	// Filter by minimum PRS score
	filtered := make([]PlacementOpportunity, 0)
	for _, opp := range opportunities {
		if opp.PRSScore >= minPRS {
			filtered = append(filtered, opp)
		}
	}

	c.JSON(http.StatusOK, gin.H{
		"opportunities": filtered,
		"total_count":   len(filtered),
		"filters": gin.H{
			"title_id": titleID,
			"min_prs":  minPRS,
		},
	})
}

// GetOpportunity handles GET /opportunities/:id
func (h *PlacementHandler) GetOpportunity(c *gin.Context) {
	id := c.Param("id")

	logrus.WithField("opportunity_id", id).Info("Getting placement opportunity")

	// TODO: Implement actual database lookup
	opportunity := PlacementOpportunity{
		ID:          id,
		TitleID:     "title_001",
		ShotID:      "shot_001",
		StartTime:   5.2,
		EndTime:     12.8,
		PRSScore:    87.5,
		SurfaceType: "wall",
		CreatedAt:   "2024-01-15T10:30:00Z",
	}

	c.JSON(http.StatusOK, opportunity)
}

// BookPlacement handles POST /bookings
func (h *PlacementHandler) BookPlacement(c *gin.Context) {
	var booking struct {
		SurfaceID     string  `json:"surface_id" binding:"required"`
		AdvertiserID  string  `json:"advertiser_id" binding:"required"`
		CampaignID    string  `json:"campaign_id" binding:"required"`
		BidAmountCPM  float64 `json:"bid_amount_cpm" binding:"required"`
		MaxImpressions int    `json:"max_impressions"`
		MinPRSScore   float64 `json:"min_prs_score"`
	}

	if err := c.ShouldBindJSON(&booking); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	logrus.WithFields(logrus.Fields{
		"surface_id":    booking.SurfaceID,
		"advertiser_id": booking.AdvertiserID,
		"campaign_id":   booking.CampaignID,
		"bid_cpm":       booking.BidAmountCPM,
	}).Info("Booking placement")

	// Create booking data map
	bookingData := map[string]interface{}{
		"surface_id":      booking.SurfaceID,
		"advertiser_id":   booking.AdvertiserID,
		"campaign_id":     booking.CampaignID,
		"bid_amount_cpm":  booking.BidAmountCPM,
		"max_impressions": booking.MaxImpressions,
		"min_prs_score":   booking.MinPRSScore,
	}

	bookingID, err := h.db.CreatePlacementBooking(bookingData)
	if err != nil {
		logrus.WithError(err).Error("Failed to create placement booking")
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to create booking"})
		return
	}

	c.JSON(http.StatusCreated, gin.H{
		"booking_id":            bookingID,
		"status":                "confirmed",
		"message":               "Placement booked successfully",
		"confirmation_time":     "2024-01-15T10:35:00Z",
		"final_cmp_rate":        booking.BidAmountCPM,
		"estimated_impressions": booking.MaxImpressions,
	})
}

// GetBooking handles GET /bookings/:id
func (h *PlacementHandler) GetBooking(c *gin.Context) {
	id := c.Param("id")

	logrus.WithField("booking_id", id).Info("Getting booking status")

	// TODO: Implement actual database lookup
	c.JSON(http.StatusOK, gin.H{
		"booking_id":            id,
		"status":               "active",
		"placement_id":         "surface_001",
		"confirmation_time":    "2024-01-15T10:35:00Z",
		"final_cpm_rate":       5.50,
		"estimated_impressions": 1000,
		"actual_impressions":    847,
	})
}

// CancelBooking handles DELETE /bookings/:id
func (h *PlacementHandler) CancelBooking(c *gin.Context) {
	id := c.Param("id")

	logrus.WithField("booking_id", id).Info("Cancelling booking")

	// TODO: Implement actual cancellation logic
	c.JSON(http.StatusOK, gin.H{
		"success":      true,
		"message":      "Booking cancelled successfully",
		"cancelled_at": "2024-01-15T11:00:00Z",
	})
}

// RecordExposure handles POST /events/exposure
func (h *PlacementHandler) RecordExposure(c *gin.Context) {
	var exposure struct {
		BookingID        string  `json:"booking_id" binding:"required"`
		ViewerID         string  `json:"viewer_id" binding:"required"`
		ExposureDuration float64 `json:"exposure_duration" binding:"required"`
		ScreenCoverage   float64 `json:"screen_coverage"`
		AttentionScore   float64 `json:"attention_score"`
	}

	if err := c.ShouldBindJSON(&exposure); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	logrus.WithFields(logrus.Fields{
		"booking_id":        exposure.BookingID,
		"exposure_duration": exposure.ExposureDuration,
		"screen_coverage":   exposure.ScreenCoverage,
	}).Info("Recording exposure event")

	// TODO: Implement actual event recording
	eventID := "event_" + exposure.BookingID + "_001"

	c.JSON(http.StatusCreated, gin.H{
		"success":  true,
		"event_id": eventID,
		"message":  "Exposure recorded successfully",
	})
}

// BatchRecordExposures handles POST /events/exposure/batch
func (h *PlacementHandler) BatchRecordExposures(c *gin.Context) {
	var batch struct {
		Events []map[string]interface{} `json:"events" binding:"required"`
	}

	if err := c.ShouldBindJSON(&batch); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	logrus.WithField("event_count", len(batch.Events)).Info("Recording batch exposure events")

	// TODO: Implement actual batch processing
	c.JSON(http.StatusCreated, gin.H{
		"processed_count": len(batch.Events),
		"failed_count":    0,
		"message":        "Batch processed successfully",
	})
}

// GetMetrics handles GET /analytics/metrics/:booking_id
func (h *PlacementHandler) GetMetrics(c *gin.Context) {
	bookingID := c.Param("booking_id")

	logrus.WithField("booking_id", bookingID).Info("Getting analytics metrics")

	// TODO: Implement actual metrics calculation
	c.JSON(http.StatusOK, gin.H{
		"booking_id":              bookingID,
		"total_impressions":       847,
		"unique_viewers":          623,
		"total_exposure_time":     4235.6,
		"average_exposure_time":   5.2,
		"average_prs_score":       89.3,
		"average_attention_score": 0.74,
		"average_screen_coverage": 23.8,
	})
}

// GetExposureEvents handles GET /analytics/events/:booking_id
func (h *PlacementHandler) GetExposureEvents(c *gin.Context) {
	bookingID := c.Param("booking_id")

	logrus.WithField("booking_id", bookingID).Info("Getting exposure events")

	// TODO: Implement actual event retrieval
	c.JSON(http.StatusOK, gin.H{
		"booking_id": bookingID,
		"events": []gin.H{
			{
				"event_id":          "event_001",
				"viewer_id":         "viewer_abc123",
				"timestamp":         "2024-01-15T10:45:00Z",
				"exposure_duration": 6.2,
				"screen_coverage":   25.4,
				"attention_score":   0.82,
			},
		},
		"total_count": 1,
	})
}